Extract duplicated code between cable and connector into BaseCable
