+++ src/main/java/com/emorn/bettercables/objects/api/forge/blocks/cable/BlockCable.java
@@ -3,40 +3,28 @@ package com.emorn.bettercables.objects.api.forge.blocks.cable;
 import com.emorn.bettercables.init.BlockInit;
 import com.emorn.bettercables.objects.api.forge.blocks.connector.BlockConnector;
 import com.emorn.bettercables.objects.api.forge.common.AxisAlignedBoundingBoxConverter;
+import com.emorn.bettercables.objects.api.forge.common.BaseCable;
 import com.emorn.bettercables.objects.application.blocks.cable.CableAxisAlignedBoundingBox;
-import com.emorn.bettercables.objects.blocks.BlockBase;
 import com.emorn.bettercables.objects.blocks.connector.NetworkManager;
 import com.emorn.bettercables.utils.IHasModel;
 import mcp.MethodsReturnNonnullByDefault;
 import net.minecraft.block.Block;
-import net.minecraft.block.SoundType;
-import net.minecraft.block.material.Material;
 import net.minecraft.block.properties.PropertyBool;
 import net.minecraft.block.state.BlockStateContainer;
 import net.minecraft.block.state.IBlockState;
-import net.minecraft.entity.Entity;
-import net.minecraft.entity.player.EntityPlayer;
-import net.minecraft.item.Item;
-import net.minecraft.item.ItemStack;
-import net.minecraft.util.EnumBlockRenderType;
 import net.minecraft.util.EnumFacing;
 import net.minecraft.util.math.AxisAlignedBB;
 import net.minecraft.util.math.BlockPos;
-import net.minecraft.util.math.RayTraceResult;
-import net.minecraft.util.math.Vec3d;
-import net.minecraft.world.Explosion;
 import net.minecraft.world.IBlockAccess;
 import net.minecraft.world.World;
 
-import javax.annotation.Nullable;
 import javax.annotation.ParametersAreNonnullByDefault;
 import java.util.ArrayList;
 import java.util.List;
-import java.util.Random;
 
 @MethodsReturnNonnullByDefault
 @ParametersAreNonnullByDefault
-public class BlockCable extends BlockBase implements IHasModel
+public class BlockCable extends BaseCable implements IHasModel
 {
     private static final PropertyBool NORTH = PropertyBool.create("north");
     private static final PropertyBool EAST = PropertyBool.create("east");
@@ -69,8 +57,7 @@ public class BlockCable extends BlockBase implements IHasModel
 
     public BlockCable(String name)
     {
-        super(name, Material.IRON);
-        setSoundType(SoundType.METAL);
+        super(name);
     }
 
     @Override
@@ -99,98 +86,6 @@ public class BlockCable extends BlockBase implements IHasModel
             .withProperty(WEST, getConnectionType(world, pos, EnumFacing.WEST))
             .withProperty(UP, getConnectionType(world, pos, EnumFacing.UP))
             .withProperty(DOWN, getConnectionType(world, pos, EnumFacing.DOWN));
-    }
-
-    @Override
-    public boolean isFullCube(IBlockState state)
-    {
-        return false;
-    }
-
-    @Override
-    public EnumBlockRenderType getRenderType(IBlockState state)
-    {
-        return EnumBlockRenderType.MODEL;
-    }
-
-    @Override
-    public AxisAlignedBB getBoundingBox(
-        IBlockState state,
-        IBlockAccess source,
-        BlockPos pos
-    )
-    {
-        EntityPlayer player = source instanceof World ? ((World) source).getClosestPlayer(
-            pos.getX(),
-            pos.getY(),
-            pos.getZ(),
-            10,
-            false
-        ) : null;
-        if (player == null) {
-            return BASE_AABB;
-        }
-
-        Vec3d start = player.getPositionEyes(1.0F);
-        Vec3d lookVec = player.getLook(1.0F);
-        Vec3d end = start.add(lookVec.scale(5));
-
-        IBlockState actualState = getActualState(state, source, pos);
-        List<AxisAlignedBB> allBoxes = retrieveAllBoxes(actualState);
-
-        AxisAlignedBB closestBox = BASE_AABB;
-        double closestDistance = Double.MAX_VALUE;
-
-        for (AxisAlignedBB box : allBoxes) {
-            RayTraceResult result = rayTrace(pos, start, end, box);
-            if (result != null) {
-                double distance = result.hitVec.distanceTo(start);
-                if (distance < closestDistance) {
-                    closestBox = box;
-                    closestDistance = distance;
-                }
-            }
-        }
-
-        return closestBox;
-    }
-
-    @Override
-    public void addCollisionBoxToList(
-        IBlockState state,
-        World worldIn,
-        BlockPos pos,
-        AxisAlignedBB entityBox,
-        List<AxisAlignedBB> collidingBoxes,
-        @Nullable Entity entityIn,
-        boolean isActualState
-    )
-    {
-        IBlockState actualState = getActualState(state, worldIn, pos);
-        List<AxisAlignedBB> allBoxes = retrieveAllBoxes(actualState);
-
-        for (AxisAlignedBB box : allBoxes) {
-            addCollisionBoxToList(pos, entityBox, collidingBoxes, box);
-        }
-    }
-
-    @Override
-    public boolean isOpaqueCube(IBlockState state)
-    {
-        return false;
-    }
-
-    @Override
-    public void onBlockDestroyedByPlayer(
-        World worldIn,
-        BlockPos pos,
-        IBlockState state
-    )
-    {
-        super.onBlockDestroyedByPlayer(worldIn, pos, state);
-        if (!worldIn.isRemote) {
-            NetworkManager.reCalculateNetworksAround(pos, worldIn);
-        }
     }
 
     @Override
@@ -242,39 +137,6 @@ public class BlockCable extends BlockBase implements IHasModel
     }
 
     @Override
-    public Item getItemDropped(
-        IBlockState state,
-        Random rand,
-        int fortune
-    )
-    {
-        return Item.getItemFromBlock(BlockInit.CABLE);
-    }
-
-    @Override
-    public void onBlockDestroyedByExplosion(
-        World worldIn,
-        BlockPos pos,
-        Explosion explosionIn
-    )
-    {
-        super.onBlockDestroyedByExplosion(worldIn, pos, explosionIn);
-        if (worldIn.isRemote) {
-            NetworkManager.reCalculateNetworksAround(pos, worldIn);
-        }
-    }
-
-    @Override
-    public ItemStack getItem(
-        World worldIn,
-        BlockPos pos,
-        IBlockState state
-    )
-    {
-        return new ItemStack(BlockInit.CABLE);
-    }
-
-    @Override
     protected BlockStateContainer createBlockState()
     {
         return new BlockStateContainer(
@@ -294,7 +156,8 @@ public class BlockCable extends BlockBase implements IHasModel
         return false;
     }
 
-    private List<AxisAlignedBB> retrieveAllBoxes(IBlockState state)
+    @Override
+    protected List<AxisAlignedBB> retrieveAllBoxes(IBlockState state)
     {
         List<AxisAlignedBB> allBoxes = new ArrayList<>();
 
@@ -325,6 +188,18 @@ public class BlockCable extends BlockBase implements IHasModel
         }
 
         return allBoxes;
+    }
+
+    @Override
+    protected AxisAlignedBB baseAABB()
+    {
+        return BASE_AABB;
+    }
+
+    @Override
+    protected Block currentBlock()
+    {
+        return BlockInit.CABLE;
     }
 
     private boolean getConnectionType(
